Update FuseWord.cpp
Simplified the reverse option by removing specifying straight the UserLangEntry. Output logs are filtered : no cross language answers (difficult as it follow a kind of "Cython" structure.
But, if it is well structured, from now on, that will be easy on the one hand to fuse two entry string and on the other hand to separate the strings instructional by language.